requirements.txt
apache-airflow
finnhub-python 
confluent-kafka 
avro-python3
snowflake-connector-python
pandas